Replaced built-in mutex with `tokio::sync::Mutex`, which is async-aware
This was a suggestion from `cargo clippy`: an async process can block a mutex, lead to data corruption,
lead to deadlocks, etc, when the mutex is not aware of the asynchronous nature of the execution scope.

This change makes the application relatively resilient to crashes, and we can even send invalid data
to some endpoints: the worker will crash, but the global state will still be usable: amazing!