Removed most panic scenarios by getting rid of `.expect()` and `.unwrap()`
Designing around this instead with higher abstractions with:

 * `Option<T>#map(Fn: T -> V) -> Option<V>`
 * `Result<E, T>#map(Fn: T -> V) -> Result<E, V>`
 * `Option<Option<T>>#flatten() -> Option<T>`

We then assemble all `Option` and `Result` types and do pattern
matching with `match (a, b, c) { (Ok(valid_a, valid_b, valid_c) => Ok(...), _ => Err(...) }`
to produce results.

This has the dis-advantage of producing a single `Err()` with not much detail about what's going
on, but perhaps it is possible to improve on the pattern matching part in future, to produce
more detailed messages.

Note also that this logic would be simplified if we wrote an actual parser for OCI data, but
that's out of reach for my current skill level: a `serde` implementation for OCI data
would be nice, but the OCI format sucks, and it is really hard to represent it cleanly in
memory, for now.