Noting why `Request::builder()` requires `.expect()` to be used
The `hyper` request builder is not a type-safe builder, because it
does not track whether it is "complete" at type level.

That's why the `.expect()` is needed: it may crash, if
misused, and it may also crash if the API changes, but
for now we don't have a better request builder.

I asked @breandan about his type-safe query builder
talk in Java / Kotlin @ https://twitter.com/Ocramius/status/1592946754772926465