Introduced `nix develop` shell, upgraded to hyper-hickori 0.8.0
This introduces `nix develop` with `rust-rover` pre-installed,
so that a reproducible environment can be used to work on the project.

The Rust toolchains are currently duplicate: only one of fenix or rust-overlay
will remain.

hyper-hickory somehow changed the suggested approach to `TokioHickoryResolver::from_default()`,
while `HickoryResolver::from_default()` is gone. We won't
go into changing that: we only adapt to the upgrade path here.

Ref: https://github.com/Gelbpunkt/hyper-hickory/commit/681fbe664bd89fe11a5b7b53fd476d613e8e8041